init.qcom.rc: Add a custom shutdown trigger to kill rmtfs
cherry-picked from upstream device/linaro/dragonboard project.

There is a kernel bug uncovered in db845c modem shutdown
path in v5.7-rc1, where "reboot" or "reboot bootloader"
commands do not work as expected when the device is
connected to a WiFi AP. The db845c hard crashes on
reboot, resulting in the device getting stuck in the usb
crash debug mode and not coming back up without a hard
power off. There is no reboot regression when the board
is not connected to any AP.

This kernel issue is already reported upstream:
https://lists.infradead.org/pipermail/ath10k/2020-April/014970.html
https://lists.infradead.org/pipermail/ath10k/2020-June/015206.html

For now here is a userspace workaround where we do not
wait for modem to shutdown gracefully (SIGTERM during
reboot), instead we stop the rmtfs service explicitly
using the "on shutdown" init trigger, thereby sending
a SIGKILL signal.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: Ieb8b65ceb979b1f698ddf7ae48ce49652b99a9dc